requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-autofixture==0.12.1
django-bootstrap-form==3.2.1
docopt==0.6.2
idna==2.5
olefile==0.44
Pillow==5.0.0
pipreqs==0.4.7
psycopg2==2.7.4
pytz==2017.2
requests==2.18.1
sorl-thumbnail==12.4a1
urllib3==1.21.1
yarg==0.1.9
